+++ src/main/java/ua/gov/court/supreme/contractwork/dto/ProjectsTotalAmounts.java
@@ -9,13 +9,13 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProjectsTotalAmounts {
-    private double totalQuantity;
+    private int totalQuantity;
     private BigDecimal totalPriceSum;
     private BigDecimal totalRemainingBalance;
     private BigDecimal totalGeneralFund;
     private BigDecimal totalSpecialFund;
 
-    public ProjectsTotalAmounts(double totalQuantity, BigDecimal totalPriceSum, BigDecimal totalGeneralFund, BigDecimal totalSpecialFund) {
+    public ProjectsTotalAmounts(int totalQuantity, BigDecimal totalPriceSum, BigDecimal totalGeneralFund, BigDecimal totalSpecialFund) {
         this.totalQuantity = totalQuantity;
         this.totalPriceSum = totalPriceSum;
         this.totalGeneralFund = totalGeneralFund;
