+++ src/main/java/ua/gov/court/supreme/contractwork/servlet/estimate/NewProjectToEstimateServlet.java
@@ -26,12 +26,10 @@ public class NewProjectToEstimateServlet extends BaseWorkServlet {
             String dkCode = req.getParameter("dkCode");
             String projectName = req.getParameter("projectName");
             String unitOfMeasure = req.getParameter("unitOfMeasure");
-            double quantity = Double.parseDouble(req.getParameter("quantity"));
+            int quantity = Integer.parseInt(req.getParameter("quantity"));
             BigDecimal price = new BigDecimal(req.getParameter("price"));
-            BigDecimal totalPrice = price.multiply(BigDecimal.valueOf(quantity)); // Recalculate for safety or take from input? Usually better to calc. User input: req.getParameter("totalPrice")
-            // Let's use user input for consistency with old logic if it was passed, but recalc is safer.
-            // Old code: double totalPrice = Double.parseDouble(req.getParameter("totalPrice"));
-            // Let's stick to parsing for now to trust frontend, but using BigDecimal.
+            BigDecimal totalPrice = price.multiply(BigDecimal.valueOf(quantity));
+            // Trust calculated total or user input? Let's use user input for consistency but parsed as BigDecimal.
             BigDecimal totalPriceInput = new BigDecimal(req.getParameter("totalPrice"));
 
             BigDecimal specialFund = new BigDecimal(req.getParameter("specialFund"));
